+++ internal/models/withdrawals.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Withdrawal struct {
-	ID          string `json:"withdrawal_id" db:"withdrawal_id"`
-	UserID      string `json:"user_id"       db:"user_id"`
-	Order       string `json:"order"         db:"order"`
-	Sum         int64  `json:"sum"           db:"sum"`
-	ProcessedAt string `json:"processed_at"  db:"processed_at"`
+	ID          string  `json:"withdrawal_id" db:"withdrawal_id"`
+	UserID      string  `json:"user_id"       db:"user_id"`
+	Order       string  `json:"order"         db:"order"`
+	Sum         float64 `json:"sum"           db:"sum"`
+	ProcessedAt string  `json:"processed_at"  db:"processed_at"`
 }
 
-func NewWithdrawal(userID string, order string, sum int64) *Withdrawal {
+func NewWithdrawal(userID string, order string, sum float64) *Withdrawal {
 	return &Withdrawal{
 		ID:     uuid.NewString(),
 		UserID: userID,
@@ -22,6 +22,6 @@ func NewWithdrawal(userID string, order string, sum int64) *Withdrawal {
 }
 
 type WithdrawalCreate struct {
-	Order string `json:"order"`
-	Sum   int64  `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float64 `json:"sum"`
 }
